Add 100% testing of the input routes files
