+++ src/views/secure/input.pug
@@ -7,7 +7,7 @@ html
 	body
 		if inIframe
 			form(action=`/input/secure/${keyPath}?token=${token}${css ? '&css=' + css : ''}`, method='post')
-				input(type='text', name='input-text')
+				input(type='text', name='input-text', value=value)
 				p#passphraseContainer turtle-wagon
 		else
 			p
